Reorg and add new home pos reward
