+++ experiments/train/solo8v2vanilla/ppo2-home-pos-mult-reward.py
@@ -5,7 +5,7 @@
 #       extension: .py
 #       format_name: percent
 #       format_version: '1.3'
-#       jupytext_version: 1.10.0
+#       jupytext_version: 1.10.1
 #   kernelspec:
 #     display_name: Python 3
 #     language: python
@@ -13,8 +13,8 @@
 # ---
 
 # %% [markdown]
-# # PPO2 on Solo8 v2 Vanilla w/ Fixed Timestamp
-# Only use the time-based stopping criteria. This is more of a rudamentary test more than anything.
+# # PPO2 on Solo8 v2 Vanilla for Quadrupedal Standing w/ a Multiplicitive Reward
+# Try to get the solo to stand on 4 feet stabley
 
 # %% [markdown]
 # ## Ensure that Tensorflow is using the GPU
@@ -30,8 +30,7 @@ else:
 # ## Define Experiment Tags
 
 # %%
-TAGS = ['solov2vanilla', 'gpu', 'standing_task', 
-        'unnormalized_actions']
+TAGS = ['solov2vanilla', 'gpu', 'home_pos_mulitiplicitive', 'unnormalized_actions']
 
 # %% [markdown]
 # # Import required libraries
@@ -71,6 +70,8 @@ config = params.WandbParameters().parse()
 config.episodes = 12500
 config.episode_length = episode_length
 
+config.target_torso_height = 0.33698 # Found experimentally
+
 config.num_workers = 6
 config.eval_frequency = 50
 config.eval_episodes = 3
@@ -94,21 +95,22 @@ config
 # - Motor encoder current values
 #
 # **Reward**
-# - How upright the TorsoIMU is. Valued in $[-1, 1]$
+# - How flat the torso is :$f$
+# - Minimize the amount of control in the joints: $c$
+# - Minimize the amount of torso movement: $m$
+# - Keeping the torso at a given height: $h$
+#
+# We'll compose the "standing" reward to be $\frac{f + h}{2}$ as $f,h \in [0, 1]$. Then, the final reward becomes:
+#
+# $$reward = \frac{f + h}{2} cm$$
+#
+# Note that since $c,m \in [0, 1]$, this enforces that $reward \in [0, 1]$
 #
 # **Termination Criteria**
 # - Terminate after $n$ timesteps
-#
-# Note that the autotrainer requires that the training environment be a `VecEnv` and the testing environment be a standard `gym.Env` for multi-processing.
-#
-# For us personally, we find that the easiest way to handle this is to create a Stable Baselines `VecEnv` generator (example can be found [here](https://stable-baselines.readthedocs.io/en/master/guide/examples.html#multiprocessing-unleashing-the-power-of-vectorized-environments)) and use that to generate both the training and testing environments.
-#
-# We also like to link our generator with our W&B config so that we can dynamically change the environments based from the web interface. 
-#
-# A full example can be seen below:
 
 # %%
-def make_env(length):
+def make_env(length, quad_standing_height):
     def _init():
         env_config = solo8v2vanilla.Solo8VanillaConfig()
         env = gym.make('solo8vanilla-v0', config=env_config, 
@@ -116,9 +118,21 @@ def make_env(length):
 
         env.obs_factory.register_observation(obs.TorsoIMU(env.robot))
         env.obs_factory.register_observation(obs.MotorEncoder(env.robot))
-
-        env.reward_factory.register_reward(1, rewards.UprightReward(env.robot))
         env.termination_factory.register_termination(terms.TimeBasedTermination(length))
+        
+        stand_reward = rewards.AdditiveReward()
+        stand_reward.client = env.client
+        
+        stand_reward.add_term(0.5, rewards.FlatTorsoReward(env.robot))
+        stand_reward.add_term(0.5, rewards.TorsoHeightReward(env.robot, quad_standing_height))
+        
+        home_pos_reward = rewards.MultiplicitiveReward(1, stand_reward,
+                                                          rewards.SmallControlReward(env.robot),
+                                                          rewards.HorizontalMoveSpeedReward(env.robot, 0))
+        home_pos_reward.client = env.client
+        
+        env.reward_factory.register_reward(1, home_pos_reward)
+
         return env
     return _init
 
@@ -135,11 +149,12 @@ from stable_baselines.common.vec_env import VecNormalize
 # Create training & testing environments
 
 # %%
-train_env = SubprocVecEnv([make_env(config.episode_length) 
+train_env = SubprocVecEnv([make_env(config.episode_length, 
+                                    config.target_torso_height) 
                            for _ in range(config.num_workers)])
-# train_env = VecNormalize(train_env, clip_reward = 1.)
 
-test_env = make_env(config.episode_length)()
+test_env = make_env(config.episode_length, 
+                    config.target_torso_height)()
 
 # %% [markdown]
 # ## Learning
